services/tesseract_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pytesseract
Pillow
python-multipart
